moves InstrumentRdrc to instrument.py
