+++ src/aind_slims_api/models/__init__.py
@@ -4,7 +4,7 @@
 
 from aind_slims_api.models.attachment import SlimsAttachment
 from aind_slims_api.models.behavior_session import SlimsBehaviorSession
-from aind_slims_api.models.instrument import SlimsInstrument
+from aind_slims_api.models.instrument import SlimsInstrument, SlimsInstrumentRdrc
 from aind_slims_api.models.metadata import SlimsMetadataReference
 from aind_slims_api.models.mouse import SlimsMouseContent
 from aind_slims_api.models.unit import SlimsUnit
@@ -16,6 +16,7 @@ __all__ = [
     "SlimsAttachment",
     "SlimsBehaviorSession",
     "SlimsInstrument",
+    "SlimsInstrumentRdrc",
     "SlimsMouseContent",
     "SlimsUnit",
     "SlimsUser",
